Factored curie and concept selector into component
